buissnes_agent/requirements.txt
mcp[cli]
uvicorn
langchain
langchain-core
langchain-openai
langchain-qdrant
langchain-text-splitters
langchain-experimental
langchain-community
qdrant-client
python-dotenv
httpx
pypdf
pandas
openpyxl
python-docx
tiktoken
scipy